Changing cat and dog to animal for frog
